backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
cloudinary==1.36.0
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.7
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0
numpy==1.24.4
face-recognition-models==0.3.0
face-recognition==1.3.0
dlib-bin==20.0.1
